convert to c++
