+++ installer/installer.cc
