+++ release/files_common
@@ -199,10 +199,12 @@ selfdrive/common/visionbuf_cl.c
 selfdrive/common/ipc.[c,h]
 selfdrive/common/swaglog.h
 selfdrive/common/swaglog.cc
-selfdrive/common/util.[c,h]
+selfdrive/common/util.cc
+selfdrive/common/util.h
 selfdrive/common/efd.[c,h]
 selfdrive/common/cqueue.[c,h]
-selfdrive/common/clutil.[c,h]
+selfdrive/common/clutil.cc
+selfdrive/common/clutil.h
 selfdrive/common/params.h
 selfdrive/common/params.cc
 selfdrive/common/mutex.h
@@ -210,7 +212,6 @@ selfdrive/common/mutex.h
 selfdrive/common/modeldata.h
 selfdrive/common/mat.h
 selfdrive/common/timing.h
-selfdrive/common/utilpp.h
 
 selfdrive/common/visionbuf.h
 selfdrive/common/visionbuf_ion.c
@@ -373,7 +374,7 @@ selfdrive/camerad/cameras/camera_qcom.h
 selfdrive/camerad/cameras/debayer.cl
 selfdrive/camerad/cameras/sensor_i2c.h
 
-selfdrive/camerad/transforms/rgb_to_yuv.c
+selfdrive/camerad/transforms/rgb_to_yuv.cc
 selfdrive/camerad/transforms/rgb_to_yuv.h
 selfdrive/camerad/transforms/rgb_to_yuv.cl
 selfdrive/camerad/transforms/rgb_to_yuv_test.cc
@@ -397,7 +398,8 @@ selfdrive/modeld/models/driving.h
 selfdrive/modeld/models/dmonitoring.cc
 selfdrive/modeld/models/dmonitoring.h
 
-selfdrive/modeld/transforms/loadyuv.[c,h]
+selfdrive/modeld/transforms/loadyuv.cc
+selfdrive/modeld/transforms/loadyuv.h
 selfdrive/modeld/transforms/loadyuv.cl
 selfdrive/modeld/transforms/transform.cc
 selfdrive/modeld/transforms/transform.h
