+++ selfdrive/camerad/cameras/camera_qcom.h
@@ -15,7 +15,7 @@
 #include "common/mat.h"
 #include "common/visionbuf.h"
 #include "common/buffering.h"
-#include "common/utilpp.h"
+#include "common/util.h"
 
 #include "camera_common.h"
