+++ selfdrive/camerad/transforms/rgb_to_yuv.h
@@ -1,5 +1,4 @@
-#ifndef RGB_TO_YUV_H
-#define RGB_TO_YUV_H
+#pragma once
 
 #include <inttypes.h>
 #include <stdbool.h>
@@ -8,10 +7,6 @@
 #include <OpenCL/cl.h>
 #else
 #include <CL/cl.h>
-#endif
-
-#ifdef __cplusplus
-extern "C" {
 #endif
 
 typedef struct {
@@ -24,9 +19,3 @@ void rgb_to_yuv_init(RGBToYUVState* s, cl_context ctx, cl_device_id device_id, i
 void rgb_to_yuv_destroy(RGBToYUVState* s);
 
 void rgb_to_yuv_queue(RGBToYUVState* s, cl_command_queue q, cl_mem rgb_cl, cl_mem yuv_cl);
-
-#ifdef __cplusplus
-}
-#endif
-
-#endif  // RGB_TO_YUV_H
