+++ selfdrive/common/SConscript
@@ -5,14 +5,14 @@ if SHARED:
 else:
   fxn = env.Library
 
-common_libs = ['params.cc', 'swaglog.cc', 'util.c', 'cqueue.c', 'gpio.cc', 'i2c.cc']
+common_libs = ['params.cc', 'swaglog.cc', 'util.cc', 'cqueue.c', 'gpio.cc', 'i2c.cc']
 
 _common = fxn('common', common_libs, LIBS="json11")
 _visionipc = fxn('visionipc', ['visionipc.c', 'ipc.c'])
 
 files = [
   'buffering.c',
-  'clutil.c',
+  'clutil.cc',
   'efd.c',
   'glutil.c',
   'visionimg.cc',
