+++ selfdrive/common/util.h
@@ -1,34 +1,20 @@
-#ifndef COMMON_UTIL_H
-#define COMMON_UTIL_H
-
+#pragma once
 #include <stdio.h>
+#include <cstdio>
+#include <unistd.h>
 
+#include <string>
+#include <memory>
+#include <sstream>
+#include <fstream>
 #ifndef sighandler_t
 typedef void (*sighandler_t)(int sig);
-#endif
-
-#ifndef __cplusplus
-
-#define min(a,b) \
- ({ __typeof__ (a) _a = (a); \
-     __typeof__ (b) _b = (b); \
-   _a < _b ? _a : _b; })
-
-#define max(a,b) \
- ({ __typeof__ (a) _a = (a); \
-     __typeof__ (b) _b = (b); \
-   _a > _b ? _a : _b; })
-
 #endif
 
 #define ARRAYSIZE(x) (sizeof(x)/sizeof(x[0]))
 
 #undef ALIGN
 #define ALIGN(x, align) (((x) + (align)-1) & ~((align)-1))
-
-#ifdef __cplusplus
-extern "C" {
-#endif
 
 // Reads a file into a newly allocated buffer.
 //
@@ -42,8 +28,70 @@ void set_thread_name(const char* name);
 int set_realtime_priority(int level);
 int set_core_affinity(int core);
 
-#ifdef __cplusplus
-}
-#endif
+namespace util {
 
-#endif
+inline bool starts_with(std::string s, std::string prefix) {
+  return s.compare(0, prefix.size(), prefix) == 0;
+}
+
+template<typename ... Args>
+inline std::string string_format( const std::string& format, Args ... args ) {
+    size_t size = snprintf( nullptr, 0, format.c_str(), args ... ) + 1;
+    std::unique_ptr<char[]> buf( new char[ size ] );
+    snprintf( buf.get(), size, format.c_str(), args ... );
+    return std::string( buf.get(), buf.get() + size - 1 );
+}
+
+inline std::string read_file(std::string fn) {
+  std::ifstream t(fn);
+  std::stringstream buffer;
+  buffer << t.rdbuf();
+  return buffer.str();
+}
+
+inline std::string tohex(const uint8_t* buf, size_t buf_size) {
+  std::unique_ptr<char[]> hexbuf(new char[buf_size*2+1]);
+  for (size_t i=0; i < buf_size; i++) {
+    sprintf(&hexbuf[i*2], "%02x", buf[i]);
+  }
+  hexbuf[buf_size*2] = 0;
+  return std::string(hexbuf.get(), hexbuf.get() + buf_size*2);
+}
+
+inline std::string base_name(std::string const & path) {
+  size_t pos = path.find_last_of("/");
+  if (pos == std::string::npos) return path;
+  return path.substr(pos + 1);
+}
+
+inline std::string dir_name(std::string const & path) {
+  size_t pos = path.find_last_of("/");
+  if (pos == std::string::npos) return "";
+  return path.substr(0, pos);
+}
+
+inline std::string readlink(std::string path) {
+  char buff[4096];
+  ssize_t len = ::readlink(path.c_str(), buff, sizeof(buff)-1);
+  if (len != -1) {
+    buff[len] = '\0';
+    return std::string(buff);
+  }
+  return "";
+}
+
+}
+
+struct unique_fd {
+  unique_fd(int fd = -1) : fd_(fd) {}
+  unique_fd& operator=(unique_fd&& uf) {
+    fd_ = uf.fd_;
+    uf.fd_ = -1;
+    return *this;
+  }
+  ~unique_fd() {
+    if (fd_ != -1) close(fd_);
+  }
+  operator int() const { return fd_; }
+  int fd_;
+};
