+++ selfdrive/loggerd/raw_logger.cc
@@ -16,7 +16,7 @@ extern "C" {
 }
 
 #include "common/swaglog.h"
-#include "common/utilpp.h"
+#include "common/util.h"
 
 #include "raw_logger.h"
