+++ selfdrive/proclogd/proclogd.cc
@@ -16,7 +16,7 @@
 #include "messaging.hpp"
 
 #include "common/timing.h"
-#include "common/utilpp.h"
+#include "common/util.h"
 
 namespace {
