+++ selfdrive/ui/qt/offroad/settings.cc
@@ -16,7 +16,7 @@
 #include "widgets/offroad_alerts.hpp"
 
 #include "common/params.h"
-#include "common/utilpp.h"
+#include "common/util.h"
 
 
 ParamsToggle::ParamsToggle(QString param, QString title, QString description, QString icon_path, QWidget *parent): QFrame(parent) , param(param) {
